Chg to read SalishSeaCast mesh mask path from YAML
It was previously hard-coded to an ERDDAP URL, which is the preferred way to do
it, but doesn't work on graham compute nodes that have not access to the
Internet. So, adding the option to enable mesh mask to be read from a clone of
the SalishSeaCast-grid repo.

This change also excludes the time coordinate from the tmask variable via array
indexing rather than .isel() so that we don't have to know the name of the time
variable because it differs between file and ERDDAP data sources.